done for gameHistoryRef
